apollo-server-hapi: enable noImplicitAny and strictNullChecks
As in #5177, the easiest way is to inline hapiApollo (AS1 API) into
ApolloServer.ts.

Also remove some unnecessary `await`s.

Fixes #5170.
